+++ apps/backend/src/modules/database/group.service.ts
@@ -35,9 +35,8 @@ export class DBGroupService {
 			select: {
 				id: true,
 				label: true,
-				createdAt: true,
 				inlinePolicies: {
-					select: { id: true, label: true, library: true, configuration: true, createdAt: true },
+					select: { id: true, label: true, library: true, configuration: true },
 				},
 			},
 		});