+++ apps/backend/src/modules/database/inline-policy.service.ts
@@ -36,10 +36,9 @@ export class DBInlinePolicyService {
 	}
 
 	public async addRule(inlinePolicyId: string, rule: Record<string, unknown>) {
-		const inlinePolicyDB = await this.prisma.inlinePolicy.findFirst({
+		const inlinePolicyDB = await this.prisma.inlinePolicy.findUniqueOrThrow({
 			where: { id: inlinePolicyId },
 			select: { rules: true },
-			rejectOnNotFound: true,
 		});
 
 		let newInlinePolicyRules: Prisma.JsonObject;
@@ -60,10 +59,9 @@ export class DBInlinePolicyService {
 	}
 
 	public async removeRule(inlinePolicyId: string, ruleName: string) {
-		const inlinePolicyDB = await this.prisma.inlinePolicy.findFirst({
+		const inlinePolicyDB = await this.prisma.inlinePolicy.findUniqueOrThrow({
 			where: { id: inlinePolicyId },
 			select: { rules: true },
-			rejectOnNotFound: true,
 		});
 
 		if (!inlinePolicyDB.rules) {
@@ -78,6 +76,13 @@ export class DBInlinePolicyService {
 		await this.prisma.inlinePolicy.update({
 			where: { id: inlinePolicyId },
 			data: { rules: rulesWithoutRule },
+		});
+	}
+
+	public getConfiguration(inlinePolicyId: string) {
+		return this.prisma.inlinePolicy.findFirst({
+			where: { id: inlinePolicyId },
+			select: { configuration: true },
 		});
 	}
 }