+++ apps/backend/src/modules/user/modules/groups/interfaces/user-group.ts
@@ -1,0 +1,7 @@
+import type { Group, InlinePolicy } from '@prisma/client';
+
+export type IUserGroupInlinePolicy = Pick<InlinePolicy, 'id' | 'label' | 'library'> & { rulesCount: number };
+
+export interface IUserGroupGetAll extends Pick<Group, 'id' | 'label'> {
+	inlinePolicies: IUserGroupInlinePolicy[];
+}