+++ apps/backend/src/modules/user/modules/inline-policies/classes/add-rule.dto.ts
@@ -1,6 +1,8 @@
+import { ApiProperty } from '@nestjs/swagger';
 import { IsJSON } from 'class-validator';
 
 export class AddRuleDto {
+	@ApiProperty({ type: String, description: 'Stringified rules object' })
 	@IsJSON()
 	readonly rule!: string;
 }