+++ apps/backend/src/modules/user/modules/inline-policies/interfaces/responses.ts
@@ -1,3 +1,9 @@
+import type { Prisma } from '@prisma/client';
+
 export interface ICreateInlinePolicy {
 	policyId: string;
+}
+
+export interface IGetConfigurationResponse {
+	configuration: Prisma.JsonValue;
 }