+++ apps/backend/src/modules/user/modules/inline-policies/update-configuration.controller.ts
@@ -1,19 +1,23 @@
-import { Body, Controller, HttpCode, HttpStatus, Logger, Param, Post, UseGuards } from '@nestjs/common';
+import { Body, Controller, HttpCode, HttpStatus, Logger, Param, Patch, UseGuards } from '@nestjs/common';
 import { CommandBus } from '@nestjs/cqrs';
+import { ApiBearerAuth, ApiOperation, ApiTags } from '@nestjs/swagger';
 
 import Routes from './inline-policies.routes';
 import { BelongingInlinePolicyGuard } from './guards/belonging-inline-policy.guard';
 import { UpdateConfigurationDto } from './classes/update-configuration.dto';
 import { UpdateConfigurationContract } from './commands/contracts/update-configuration.contract';
 
+@ApiTags('Inline Policies')
 @Controller(Routes.CONTROLLER)
 export class UpdateConfigurationController {
 	private readonly logger = new Logger(UpdateConfigurationController.name);
 
 	constructor(private readonly commandBus: CommandBus) {}
 
+	@ApiOperation({ description: 'Update the configuration of a policy by its identifier' })
+	@ApiBearerAuth('access-token')
 	@UseGuards(BelongingInlinePolicyGuard)
-	@Post(Routes.UPDATE_CONFIGURATION)
+	@Patch(Routes.UPDATE_CONFIGURATION)
 	@HttpCode(HttpStatus.OK)
 	public async updateConfiguration(
 		@Param('policy_id') policyId: string,